Fix for bad transaction handling on SMART server
Uploading test patients to the SMART server was causing errors because
it complained we were updating conditions associated w/ deleted
patients (even though the same transaction was re-adding the deleted
patient).

This PR changes the approach to upload the patient resource first, and
then upload the other resources (if patient upload succeeded).