Fix concurrent conditions & encounter dx
The prior code's logic for sub-sections w/ mutliple tables had an all-or-nothing strategy (if *any* item didn't have the trigger field, then *no* items were rendered).  This caused problems for the sub-section showing conditions and encounter dx.

This new code uses changes 'trigger' to 'filter' and uses it to filter the items shown in a table.  In order to ensure no items are lost, however, developers need to ensure that the filters in a sub-section cover all possibilities (and likely also want them to be mutually exclusive).