Reference sort icons via relative imports
When the sort icons are referenced via absolute paths, they are broken when deploying to subfolders.  Instead, we need to import them via a relative path.  Because of the way the CSS build works, we can't import things outside of "src", so this also required moving the icons to be under "src".