Added ability to POST application analytics based on configuration. Fixed broken test.
Fixed a test previously broken due to a new flag added.
Added the ability to post application analytics, which include a time and flag for whether the inclusion criteria were met.
If inclusion criteria was met, analyics also includes the total number of flagged entries and each section, broken down into subsections, with a total number of flagged entries on each subsection.