Uneject and Upgrade React-Scripts
This commit makes the Pain Management Summary a normal Create React App
/ React Scripts app again.  In order to leverage the latest fhirclient,
we need an updated version of react-scripts.  Since PMSA was ejected,
updating react-scripts was not trivial in the ejected state.  Since we
initially created this app, CRA has improved a lot and other approaches
to customizing configuration have emerged.  That being the case, we
decided to make PMSA a standard CRA app again, allowing for easier
upgrades in the future as well.

To accomplish this, I used CRA to create a brand new app and then
moved the PMSA code into the new CRA app, making adjustments/fixes
as necessary.  I also brought in "rescripts" to assist in customizing the
webpack config as necessary.  I then applied all these changes to the
existing repo, resulting in this commit.

Note that I also changed the approach to how the static app is built and
served since we no longer control the build script.  Now it will be
built at the root of the build folder, but served at the
AHRQ-CDS-Connect-PAIN-MANAGEMENT-SUMMARY path, as required
by the current configuration.

In theory, this version of PMSA should function identically to the
previous ejected version.  There was no attempt to add new
functionality (aside from what CRA gives you).