FHIR 4.0.1 Update
Updates to the Pain Management Summary to keep up w/ standards and libraries.

* Update CQL, CQL exec libraries, translations, and FHIR version
* Use latest CQL execution libraries (based on CQL 1.4/1.5)
* Add support for VSAC URL-based value set identifiers
* Translate CQL using CQL-to-ELM 1.5.2 translator
* Update CQL to latest FHIR 4.0.1 version
* Script to update date fields in test patients (plus test patient updates)
* Bump .node-version to 12.20.0 and fix typo

This code was previously reviewed and approved by Noranda Brown, but was put on hold until closer to the FHIR 4.0.1 update in the CDS Repository. Since then, it has been rebased, recompiled, and value sets / patients automatically updated.